onb: flow updated with test cases
