+++ lib/common/screens/splash_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:hirevire_app/user_interface/controllers/navigation_controller.dart';
+import 'package:hirevire_app/common/controllers/common_nav_controller.dart';
 import 'package:hirevire_app/common/widgets/custom_image_view.dart';
 import 'package:hirevire_app/constants/color_constants.dart';
 import 'package:hirevire_app/constants/image_constants.dart';
@@ -9,8 +9,8 @@ import 'package:hirevire_app/utils/size_util.dart';
 class SplashScreen extends StatelessWidget {
   SplashScreen({super.key});
 
-  final NavigationController navController =
-      Get.put(tag: 'navController', NavigationController());
+  final CommonNavController commonNavController =
+      Get.put(tag: 'commonNavController', CommonNavController());
 
   @override
   Widget build(BuildContext context) {
