+++ lib/common/widgets/body_text_widget.dart
@@ -5,14 +5,16 @@ class BodyTextWidget extends StatelessWidget {
   const BodyTextWidget({
     super.key,
     required this.text,
+    this.style,
   });
   final String text;
+  final TextStyle? style;
 
   @override
   Widget build(BuildContext context) {
     return Text(
       text,
-      style: AppTextThemes.bodyTextStyle(context),
+      style: style ?? AppTextThemes.bodyTextStyle(context),
     );
   }
 }
