+++ lib/constants/endpoint_constants.dart
@@ -3,6 +3,7 @@ class Endpoints {
   static const String searchJobTitles = 'master/jobTitles/by-name';
   static const String searchCompanies = 'master/companies/by-name';
   static const String searchSkills = 'master/skills/by-name';
+  static const String getTexts = 'content';
 
   // POST Endpoints
   static const String sendOtp = 'auth/sendOTP';
