+++ lib/constants/persistence_keys.dart
@@ -1,11 +1,17 @@
 class PersistenceKeys {
   // ------------- Common keys -------------
 
+  static const String isNew = 'isNew';
+  static const String isEmployee = 'isEmployee';
+  static const String isSignedIn = 'isSignedIn';
   static const String authToken = 'authToken';
-  static const String isSignedIn = 'is_signed_in';
-  static const String isCompany = 'is_therapist';
+  static const String name = 'name';
+  static const String email = 'email';
 
-  // ------------- Company keys -------------
+  // ------------- Employee keys -------------
 
   // ------------- User keys -------------
+  static const String isEmailVerified = 'isEmailVerified';
+  static const String isUserRegistered = 'isUserRegistered';
+  static const String isProfileComplete = 'isProfileComplete';
 }
