+++ lib/employer_interface/controllers/emp_onb_controller.dart
@@ -1,0 +1,3 @@
+import 'package:get/get.dart';
+
+class EmpOnbController extends GetxController {}
