+++ lib/employer_interface/presentation/auth/emp_login_screen.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:hirevire_app/employer_interface/controllers/emp_onb_controller.dart';
+
+class EmpLoginScreen extends GetWidget<EmpOnbController> {
+  const EmpLoginScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold();
+  }
+}
