+++ lib/user_interface/controllers/user_onb_controller.dart
@@ -1,96 +1,38 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:get/get.dart';
+import 'package:hirevire_app/common/models/text_model.dart';
 import 'package:hirevire_app/constants/endpoint_constants.dart';
 import 'package:hirevire_app/constants/error_constants.dart';
-import 'package:hirevire_app/constants/global_constants.dart';
+import 'package:hirevire_app/constants/persistence_keys.dart';
 import 'package:hirevire_app/services/api_service.dart';
-import 'package:hirevire_app/user_interface/presentation/onboarding/professional_details/bio_section.dart';
-import 'package:hirevire_app/user_interface/presentation/onboarding/personal_details.dart/name_dob_section.dart';
-import 'package:hirevire_app/user_interface/presentation/onboarding/personal_details.dart/phone_number.dart';
-import 'package:hirevire_app/user_interface/presentation/onboarding/professional_details/experience_section.dart';
-import 'package:hirevire_app/user_interface/presentation/onboarding/professional_details/location_section.dart';
-import 'package:hirevire_app/user_interface/presentation/onboarding/professional_details/skills_section.dart';
-import 'package:hirevire_app/user_interface/presentation/onboarding/professional_details/social_urls_section.dart';
-import 'package:hirevire_app/user_interface/routes/app_routes.dart';
+import 'package:hirevire_app/routes/app_routes.dart';
 import 'package:hirevire_app/utils/datetime_util.dart';
 import 'package:hirevire_app/utils/log_handler.dart';
+import 'package:hirevire_app/utils/persistence_handler.dart';
 import 'package:hirevire_app/utils/show_toast_util.dart';
 import 'package:hirevire_app/utils/validation_util.dart';
-import 'package:image_picker/image_picker.dart';
 
 class UserOnbController extends GetxController {
   late ApiClient apiClient;
-  String country = "IN";
-  String countryCode = "+91";
-  String defaultItemId = "Other";
 
   RxBool isLoading = false.obs;
   RxBool isEmailValid = false.obs;
   RxBool isOtpLengthValid = false.obs;
-  RxInt currentProgressIndex = 0.obs;
   RxBool isDobValid = false.obs;
+  RxBool basicDetailsValid = false.obs;
   RxString errorMsg = ''.obs;
-  RxBool isStep1Valid = false.obs;
-  RxBool isNumberValid = false.obs;
-  RxBool isHeadlineValid = false.obs;
-  RxBool skillsAdded = false.obs;
-  RxBool addingExp = false.obs;
-  RxBool expAdded = false.obs;
   RxBool isSigningUp = false.obs;
-  RxString skillsSearchQuery = ''.obs;
-  RxString titleSearchQuery = ''.obs;
-  RxString companySearchQuery = ''.obs;
-  RxMap<String, dynamic> headlineObj = <String, dynamic>{}.obs;
-  RxMap<String, dynamic> companyJobTitleObj = <String, dynamic>{}.obs;
-  RxMap<String, dynamic> companyObj = <String, dynamic>{}.obs;
-  RxList<Map<String, dynamic>> selectedSkills = <Map<String, dynamic>>[].obs;
-  RxList<Map<String, dynamic>> filteredSkillsSuggestions =
-      <Map<String, dynamic>>[].obs;
-  RxList<Map<String, dynamic>> filteredTitleSuggestions =
-      <Map<String, dynamic>>[].obs;
-  RxList<Map<String, dynamic>> filteredCompanySuggestions =
-      <Map<String, dynamic>>[].obs;
-  RxList<Map<String, dynamic>> addedExp = <Map<String, dynamic>>[].obs;
-  RxList<Map<String, dynamic>> socialProfiles = <Map<String, dynamic>>[].obs;
-  RxString employmentType = GlobalConstants.employmentTypes[0].obs;
-  RxString socialProfileType =
-      GlobalConstants.socialProfileTypesMap.keys.first.obs;
-  RxString preferredJobMode = GlobalConstants.locationTypes[0].obs;
-  RxBool isAddingUrl = false.obs;
-  RxString locationType = GlobalConstants.locationTypes[0].obs;
-  RxBool stillWorking = true.obs;
-  RxString startMonth = GlobalConstants.monthsMap.keys.first.obs;
-  RxString startYear = GlobalConstants.years[0].obs;
-  RxString endMonth = GlobalConstants.monthsMap.keys.first.obs;
-  RxString endYear = GlobalConstants.years[0].obs;
 
-  final imagePicker = ImagePicker();
-  Rx<File?> profilePic = Rx<File?>(null);
+  RxList<TextModel> texts = <TextModel>[].obs;
 
   // Text Controllers
-
-  final PageController pageController = PageController(initialPage: 0);
   TextEditingController emailController = TextEditingController();
   final List<TextEditingController> otpControllers =
       List.generate(6, (_) => TextEditingController());
   TextEditingController nameController = TextEditingController();
   final List<TextEditingController> dobControllers =
       List.generate(3, (_) => TextEditingController());
-  TextEditingController numberController = TextEditingController();
-  TextEditingController headlineController = TextEditingController();
-  TextEditingController bioController = TextEditingController();
-  TextEditingController skillsSearchController = TextEditingController();
-  TextEditingController expTitleController = TextEditingController();
-  TextEditingController expDescriptionController = TextEditingController();
-  TextEditingController companyIdController =
-      TextEditingController(text: 'Other');
-  TextEditingController companyNameController = TextEditingController();
-  TextEditingController companyLocationController = TextEditingController();
-  TextEditingController socialUrlController = TextEditingController();
-  TextEditingController locationController = TextEditingController();
 
   // Text Focus nodes
 
@@ -98,46 +40,17 @@ class UserOnbController extends GetxController {
   final List<FocusNode> otpFocusNodes = List.generate(6, (_) => FocusNode());
   FocusNode nameFocusNode = FocusNode();
   final List<FocusNode> dobFocusNodes = List.generate(3, (_) => FocusNode());
-  FocusNode numberFocusNode = FocusNode();
-  FocusNode headlineFocusNode = FocusNode();
-  FocusNode bioFocusNode = FocusNode();
-  final FocusNode searchFocusNode = FocusNode();
-  final FocusNode expTitleFocusNode = FocusNode();
-  final FocusNode expDescriptionFocusNode = FocusNode();
-  final FocusNode companyNameFocusNode = FocusNode();
-  final FocusNode companyLocationFocusNode = FocusNode();
-  final FocusNode socialUrlFocusNode = FocusNode();
-  final FocusNode locationFocusNode = FocusNode();
 
   @override
   void onInit() {
     super.onInit();
     apiClient = ApiClient();
-    // Bind the filteredSuggestions to changes in searchQuery
-    debounce(skillsSearchQuery, (_) => suggestSkills(),
-        time: const Duration(milliseconds: 300));
-    debounce(titleSearchQuery, (_) => suggestTitles(),
-        time: const Duration(milliseconds: 300));
-    debounce(companySearchQuery, (_) => suggestCompany(),
-        time: const Duration(milliseconds: 300));
   }
 
   // ----------------- Methods -----------------
 
   showBottomToast(String message) {
     ToastWidgit.bottomToast(message);
-  }
-
-  List<Widget> sectionWidgets() {
-    return [
-      const NameDobSection(),
-      const NumberSection(),
-      const BioSection(),
-      const SkillsSection(),
-      const ExperienceSection(),
-      const SocialUrlsSection(),
-      const LocationSection(),
-    ];
   }
 
   void validateEmail({String? email}) {
@@ -195,7 +108,14 @@ class UserOnbController extends GetxController {
       Map<String, dynamic> response = await apiClient.post(endpoint, body);
 
       if (response['success']) {
-        navigateToIntroScreen();
+        PersistenceHandler.setBool(PersistenceKeys.isNew, false);
+
+        bool accountExist = response['body']['accountExist'];
+        if (accountExist) {
+          navigateToBaseNav();
+        } else {
+          navigateToNameScreen();
+        }
       } else {
         String errorMsg =
             response['error']['message'] ?? Errors.somethingWentWrong;
@@ -210,45 +130,10 @@ class UserOnbController extends GetxController {
     }
   }
 
-  Future<void> suggestTitles() async {
-    if (titleSearchQuery.isEmpty) {
-      filteredTitleSuggestions.value = [];
-    } else {
-      String endpoint = Endpoints.searchJobTitles;
-      String searchQuery = "$endpoint/${titleSearchQuery.value}";
-
-      try {
-        Map<String, dynamic> response = await apiClient.get(searchQuery);
-        if (response['success']) {
-          List<Map<String, dynamic>> convertedList =
-              List<Map<String, dynamic>>.from(response['body']['data']);
-          filteredTitleSuggestions.value = convertedList;
-        } else {
-          String errorMsg =
-              response['error']['message'] ?? Errors.somethingWentWrong;
-          showBottomToast(errorMsg);
-          LogHandler.error(errorMsg);
-        }
-      } catch (error) {
-        showBottomToast(error.toString());
-        LogHandler.error(error);
-      }
-    }
-  }
-
-  void setTitle(Map<String, dynamic> selectedTitle) {
-    headlineController.text = selectedTitle['name'];
-    titleSearchQuery.value = '';
-    headlineObj.value = {
-      'data': selectedTitle['_id'],
-      'name': selectedTitle['name'],
-    };
-  }
-
   validateName() {
     bool isNameValid = nameController.value.text.trim().isNotEmpty;
     if (!isNameValid) {
-      isStep1Valid.value = false;
+      basicDetailsValid.value = false;
     }
     return isNameValid;
   }
@@ -256,29 +141,63 @@ class UserOnbController extends GetxController {
   bool validateDobValues(int index, String value) {
     int intValue = value.isNotEmpty ? int.parse(value) : 0;
     bool isValueValid = false;
+    bool shiftFocus = false;
     switch (index) {
       case 0:
-        if (value.length == 2) {
+        if (value.length <= 2) {
           isValueValid = intValue > 0 && intValue <= 31;
+          shiftFocus = isValueValid && value.length == 2;
           errorMsg.value = isValueValid ? "" : "*Invalid date";
         }
         break;
       case 1:
-        if (value.length == 2) {
+        if (intValue == 2) {
+          int date = int.parse(dobControllers[0].text);
+          isValueValid = date > 0 && date <= 29;
+          errorMsg.value = isValueValid ? "" : "*Invalid date";
+        } else if (value.length <= 2) {
           isValueValid = intValue > 0 && intValue <= 12;
           errorMsg.value = isValueValid ? "" : "*Invalid month";
         }
         break;
       case 2:
         if (value.length == 4) {
-          isValueValid =
-              intValue > 1930 && intValue <= 2015; // hardcoded for now
-          errorMsg.value = isValueValid ? "" : "*Invalid year";
+          bool isLeap = isLeapYear(intValue);
+          int validMinMaxYear = DateTime.now().year - 18;
+          int date = int.parse(dobControllers[0].text);
+          int month = int.parse(dobControllers[1].text);
+
+          if (intValue > validMinMaxYear) {
+            errorMsg.value = "*You must be over 18";
+          } else if (month == 2 && !isLeap) {
+            isValueValid = date > 0 && date <= 28;
+            errorMsg.value = isValueValid ? "" : "*Invalid date";
+          } else {
+            isValueValid =
+                intValue > 1930 && intValue <= validMinMaxYear; // min age: 18
+            errorMsg.value = isValueValid ? "" : "*Invalid year";
+          }
         }
         break;
     }
     isDataProvided();
-    return isValueValid;
+    return shiftFocus;
+  }
+
+  bool isLeapYear(int year) {
+    if (year % 4 == 0) {
+      if (year % 100 == 0) {
+        if (year % 400 == 0) {
+          return true; // Divisible by 400
+        } else {
+          return false; // Divisible by 100 but not by 400
+        }
+      } else {
+        return true; // Divisible by 4 but not by 100
+      }
+    } else {
+      return false; // Not divisible by 4
+    }
   }
 
   bool isDobCorrect() {
@@ -333,14 +252,14 @@ class UserOnbController extends GetxController {
       return text;
     }).join();
 
-    isStep1Valid.value = dob.length == 8 && validateName();
+    basicDetailsValid.value = dob.length == 8 && validateName();
   }
 
   validateNameDob() {
-    isStep1Valid.value = validateName() && isDobCorrect();
+    basicDetailsValid.value = validateName() && isDobCorrect();
   }
 
-  DateTime? getDob() {
+  DateTime getDob() {
     try {
       String day = dobControllers[0].text.trim().padLeft(2, '0');
       String month = dobControllers[1].text.trim().padLeft(2, '0');
@@ -349,319 +268,7 @@ class UserOnbController extends GetxController {
       DateTime dob = DatetimeUtil.getDateTimeFromString(day, month, year);
       return dob;
     } catch (e) {
-      return null;
-    }
-  }
-
-  void moveToNextStep() {
-    if (currentProgressIndex < sectionWidgets().length - 1) {
-      errorMsg.value = '';
-      currentProgressIndex += 1;
-      pageController.animateToPage(
-        currentProgressIndex.value,
-        duration: const Duration(milliseconds: 300),
-        curve: Curves.easeInOut,
-      );
-    } else if (currentProgressIndex.value == sectionWidgets().length - 1) {
-      // signup
-    }
-  }
-
-  validateNumber() {
-    if (numberController.value.text.trim().length == 10) {
-      // errorMsg.value = "";
-      isNumberValid.value = true;
-      numberFocusNode.unfocus();
-      moveToNextStep();
-    } else {
-      isNumberValid.value = false;
-      // errorMsg.value = "*Invalid number";
-    }
-  }
-
-  validateHeadline() {
-    if (headlineController.value.text.isNotEmpty) {
-      isHeadlineValid.value = true;
-      errorMsg.value = "";
-      headlineFocusNode.unfocus();
-      bioFocusNode.unfocus();
-      moveToNextStep();
-    } else {
-      isHeadlineValid.value = false;
-      errorMsg.value = "*No headline provided";
-    }
-  }
-
-  validateSkillsSection() {
-    if (selectedSkills.isNotEmpty) {
-      skillsAdded.value = true;
-      errorMsg.value = "";
-      moveToNextStep();
-    } else {
-      skillsAdded.value = false;
-      errorMsg.value = "No skills added";
-    }
-  }
-
-  Future<void> suggestSkills() async {
-    if (skillsSearchQuery.isEmpty) {
-      filteredSkillsSuggestions.value = [];
-    } else {
-      String endpoint = Endpoints.searchSkills;
-      String searchQuery = "$endpoint/${skillsSearchQuery.value}";
-
-      try {
-        Map<String, dynamic> response = await apiClient.get(searchQuery);
-
-        if (response['success']) {
-          List<Map<String, dynamic>> convertedList =
-              List<Map<String, dynamic>>.from(response['body']['data']);
-          filteredSkillsSuggestions.value = convertedList;
-        } else {
-          String errorMsg =
-              response['error']['message'] ?? Errors.somethingWentWrong;
-          showBottomToast(errorMsg);
-          LogHandler.error(errorMsg);
-        }
-      } catch (error) {
-        showBottomToast(error.toString());
-        LogHandler.error(error);
-      }
-    }
-  }
-
-  void addSkill(Map<String, dynamic> selectedSkill) {
-    Map<String, dynamic> skillMap = {
-      'data': selectedSkill['_id'],
-      'name': selectedSkill['name'],
-    };
-
-    bool skillExists = false;
-
-    if (selectedSkills.isEmpty) {
-      selectedSkills.add(skillMap);
-    } else {
-      for (var i = 0; i < selectedSkills.length; i++) {
-        var skill = selectedSkills[i];
-        if (skill['name'].toString().toLowerCase() ==
-            selectedSkill['name'].toString().toLowerCase()) {
-          skillExists = true;
-          return;
-        }
-      }
-      if (!skillExists) {
-        selectedSkills.add(skillMap);
-      }
-    }
-    skillsSearchController.clear();
-  }
-
-  void removeSkill(Map<String, dynamic> titleObj) {
-    selectedSkills.removeWhere((skill) => skill['name'] == titleObj['name']);
-  }
-
-  Future<void> suggestCompany() async {
-    if (companySearchQuery.isEmpty) {
-      filteredCompanySuggestions.value = [];
-    } else {
-      String endpoint = Endpoints.searchCompanies;
-      String searchQuery = "$endpoint/${companySearchQuery.value}";
-
-      try {
-        Map<String, dynamic> response = await apiClient.get(searchQuery);
-
-        if (response['success']) {
-          List<Map<String, dynamic>> convertedList =
-              List<Map<String, dynamic>>.from(response['body']['data']);
-          filteredCompanySuggestions.value = convertedList;
-        } else {
-          String errorMsg =
-              response['error']['message'] ?? Errors.somethingWentWrong;
-          showBottomToast(errorMsg);
-          LogHandler.error(errorMsg);
-        }
-      } catch (error) {
-        showBottomToast(error.toString());
-        LogHandler.error(error);
-      }
-    }
-  }
-
-  setCompanyJobTitle(Map<String, dynamic> selectedCompanyJobTitle) {
-    expTitleController.text = selectedCompanyJobTitle['name'];
-    titleSearchQuery.value = '';
-    companyJobTitleObj.value = {
-      'data': selectedCompanyJobTitle['_id'],
-      'name': selectedCompanyJobTitle['name'],
-    };
-  }
-
-  setCompanyName(Map<String, dynamic> selectedCompany) {
-    companyNameController.text = selectedCompany['name'];
-    companySearchQuery.value = '';
-    companyObj.value = {
-      'data': selectedCompany['_id'],
-      'name': selectedCompany['name'],
-    };
-  }
-
-  searchCompanyLocation() {}
-
-  toggleStillWorkingValue() {
-    stillWorking.value = !stillWorking.value;
-  }
-
-  saveExp() {
-    bool isEndDateAdded;
-
-    if (stillWorking.value) {
-      isEndDateAdded = true;
-    } else if (endMonth.value != 'Month' && endYear.value != 'Year') {
-      isEndDateAdded = true;
-    } else {
-      isEndDateAdded = false;
-    }
-
-    if (expTitleController.value.text.isNotEmpty &&
-        companyNameController.value.text.isNotEmpty &&
-        startMonth.value != 'Month' &&
-        startYear.value != 'Year' &&
-        isEndDateAdded) {
-      errorMsg.value = "";
-      Map<String, dynamic> exp = {
-        'title': companyJobTitleObj,
-        'company': companyObj,
-        "location": {
-          "country": companyLocationController.value.text.trim(),
-          "state": "",
-          "city": "",
-          "address": {"line1": "", "line2": ""}
-        },
-        'description': expDescriptionController.value.text.trim(),
-        'employmentType': employmentType.value == "Please select"
-            ? null
-            : employmentType.value,
-        'jobMode':
-            locationType.value == "Please select" ? null : locationType.value,
-        'startDate': DatetimeUtil.getDateTimeFromString(
-          "01",
-          GlobalConstants.monthsMap[startMonth.value]
-              .toString()
-              .padLeft(2, '0'),
-          startYear.value,
-        ).toIso8601String(),
-        "stillWorking": stillWorking.value,
-        'endDate': stillWorking.value
-            ? null
-            : DatetimeUtil.getDateTimeFromString(
-                "28",
-                GlobalConstants.monthsMap[endMonth.value]
-                    .toString()
-                    .padLeft(2, '0'),
-                endYear.value,
-              ).toIso8601String(),
-        "skills": [],
-        "media": [
-          {
-            "title": "Resume",
-            "url": "https://example.com/project-a",
-            "type": "image",
-            "thumbnail": "https://example.com/project-a-thumbnail"
-          }
-        ],
-      };
-      addedExp.add(exp);
-      clearExpControllers();
-    } else {
-      errorMsg.value = "*Add required fields";
-    }
-  }
-
-  validateAddedExp() {
-    if (addedExp.isNotEmpty) {
-      expAdded.value = true;
-      errorMsg.value = "";
-      clearExpControllers();
-      moveToNextStep();
-    } else {
-      expAdded.value = false;
-      errorMsg.value = "*No experience added";
-    }
-  }
-
-  void clearExpControllers() {
-    addingExp.value = false;
-    stillWorking.value = true;
-    expTitleController.clear();
-    companyNameController.clear();
-    companyLocationController.clear();
-    expDescriptionController.clear();
-    employmentType.value = GlobalConstants.employmentTypes[0];
-    locationType.value = GlobalConstants.locationTypes[0];
-    startMonth.value = GlobalConstants.monthsMap.keys.first;
-    startYear.value = GlobalConstants.years[0];
-    endMonth.value = GlobalConstants.monthsMap.keys.first;
-    endYear.value = GlobalConstants.years[0];
-  }
-
-  checkUrlStatus() {
-    isAddingUrl.value = socialUrlController.value.text.trim().isNotEmpty &&
-        socialProfileType.value.trim() != 'Select';
-  }
-
-  clearSocialControllers() {
-    socialUrlController.clear();
-    socialProfiles.clear();
-    socialProfileType.value = GlobalConstants.socialProfileTypesMap.keys.first;
-  }
-
-  addSocialUrl() {
-    checkUrlStatus();
-    if (isAddingUrl.value) {
-      socialProfiles.add({
-        "platform": socialProfileType.value,
-        "url": socialUrlController.value.text.trim()
-      });
-      socialUrlController.clear();
-      errorMsg.value = '';
-      checkUrlStatus();
-    } else {
-      if (socialUrlController.value.text.trim().isEmpty) {
-        errorMsg.value = '*URL missing';
-      } else if (socialProfileType.value.trim() == 'Select') {
-        errorMsg.value = '*Select platform type';
-      }
-    }
-  }
-
-  Future<void> pickImage() async {
-    final pickedFile = await imagePicker.pickImage(
-        source: ImageSource.gallery, imageQuality: 80);
-
-    if (pickedFile != null) {
-      profilePic.value = File(pickedFile.path);
-    }
-  }
-
-  Future<void> uploadProfilePic() async {
-    String endpoint = Endpoints.uploadProfilePicture;
-
-    try {
-      Map<String, dynamic> response =
-          await apiClient.uploadImageOrVideo(endpoint, profilePic.value!);
-
-      if (response['success']) {
-        // String url = response['body']['data'];
-        // TODO: save URL in user model
-      } else {
-        String errorMsg =
-            response['error']['message'] ?? Errors.somethingWentWrong;
-        showBottomToast(errorMsg);
-        LogHandler.error(errorMsg);
-      }
-    } catch (error) {
-      showBottomToast(error.toString());
-      LogHandler.error(errorMsg);
+      return DateTime.now();
     }
   }
 
@@ -669,45 +276,34 @@ class UserOnbController extends GetxController {
     isSigningUp.value = true;
 
     String endpoint = Endpoints.signup;
-    String? number;
-
-    if (numberController.value.text.trim().length == 10) {
-      number = "$countryCode ${numberController.value.text.trim()}";
-    }
+    String dob = getDob().toIso8601String();
+    String email = emailController.value.text.trim().isEmpty
+        ? await PersistenceHandler.getString(PersistenceKeys.email)
+        : emailController.value.text.trim();
 
     Map<String, dynamic> body = {
       "name": nameController.value.text.trim(),
-      "email": emailController.value.text.trim(),
-      "phone": number,
-      "profilePicUrl": null,
-      "headline": headlineObj,
-      "bio": bioController.value.text.trim(),
-      "skills": selectedSkills,
-      "experience": addedExp,
-      "socialUrls": socialProfiles,
-      "location": {
-        "country": locationController.value.text.trim(),
-        "state": "",
-        "city": "",
-        "address": {"line1": "", "line2": ""}
-      },
-      "preferredJobModes": [preferredJobMode.value],
+      "email": email,
+      "birthDate": dob,
     };
     LogHandler.debug(body);
 
     try {
       Map<String, dynamic> response = await apiClient.post(endpoint, body);
+      LogHandler.debug(response);
 
       if (response['success']) {
-        // TODO: save user data to user model
+        String token = response['body']['token'];
+        String name = response['body']['data']['name'];
 
-        if (profilePic.value != null) {
-          await uploadProfilePic();
-        }
+        PersistenceHandler.setString(PersistenceKeys.authToken, token);
+        PersistenceHandler.setString(PersistenceKeys.name, name);
+        PersistenceHandler.setBool(PersistenceKeys.isSignedIn, true);
+        PersistenceHandler.setBool(PersistenceKeys.isUserRegistered, true);
 
-        List<String> name = nameController.value.text.trim().split(' ');
+        List<String> nameSplit = name.split(' ');
         navigateToBaseNav();
-        showBottomToast("Welcome ${name[0]}!");
+        showBottomToast("Welcome ${nameSplit[0]}!");
       } else {
         String errorMsg =
             response['error']['message'] ?? Errors.somethingWentWrong;
@@ -728,11 +324,16 @@ class UserOnbController extends GetxController {
     Get.toNamed(AppRoutes.otpScreen);
   }
 
-  navigateToIntroScreen() {
-    Get.toNamed(AppRoutes.introScreen);
+  navigateToNameScreen() {
+    PersistenceHandler.setBool(PersistenceKeys.isEmployee, false);
+    PersistenceHandler.setBool(PersistenceKeys.isEmailVerified, true);
+    PersistenceHandler.setString(
+        PersistenceKeys.email, emailController.text.trim());
+    Get.toNamed(AppRoutes.nameScreen);
   }
 
   navigateToBaseNav() {
+    PersistenceHandler.setBool(PersistenceKeys.isSignedIn, true);
     Get.toNamed(AppRoutes.userBaseNavigator);
   }
 }
