+++ lib/user_interface/presentation/profile/personal_details.dart/phone_number.dart
@@ -8,10 +8,10 @@ import 'package:hirevire_app/common/widgets/heading_large.dart';
 import 'package:hirevire_app/common/widgets/spacing_widget.dart';
 import 'package:hirevire_app/common/widgets/text_field.dart';
 import 'package:hirevire_app/constants/image_constants.dart';
-import 'package:hirevire_app/user_interface/controllers/user_onb_controller.dart';
+import 'package:hirevire_app/user_interface/controllers/profile_controller.dart';
 import 'package:hirevire_app/utils/size_util.dart';
 
-class NumberSection extends GetWidget<UserOnbController> {
+class NumberSection extends GetWidget<ProfileController> {
   const NumberSection({super.key});
 
   @override
