+++ lib/user_interface/presentation/profile/professional_details/location_section.dart
@@ -8,10 +8,10 @@ import 'package:hirevire_app/common/widgets/loader_circular.dart';
 import 'package:hirevire_app/common/widgets/text_field.dart';
 import 'package:hirevire_app/constants/global_constants.dart';
 import 'package:hirevire_app/constants/image_constants.dart';
-import 'package:hirevire_app/user_interface/controllers/user_onb_controller.dart';
+import 'package:hirevire_app/user_interface/controllers/profile_controller.dart';
 import 'package:hirevire_app/utils/size_util.dart';
 
-class LocationSection extends GetWidget<UserOnbController> {
+class LocationSection extends GetWidget<ProfileController> {
   const LocationSection({super.key});
 
   @override
