+++ lib/user_interface/presentation/profile/professional_details/skills_section.dart
@@ -10,10 +10,10 @@ import 'package:hirevire_app/common/widgets/spacing_widget.dart';
 import 'package:hirevire_app/common/widgets/text_field.dart';
 import 'package:hirevire_app/constants/color_constants.dart';
 import 'package:hirevire_app/constants/image_constants.dart';
-import 'package:hirevire_app/user_interface/controllers/user_onb_controller.dart';
+import 'package:hirevire_app/user_interface/controllers/profile_controller.dart';
 import 'package:hirevire_app/utils/size_util.dart';
 
-class SkillsSection extends GetWidget<UserOnbController> {
+class SkillsSection extends GetWidget<ProfileController> {
   const SkillsSection({super.key});
 
   @override
