+++ lib/user_interface/presentation/profile/professional_details/social_urls_section.dart
@@ -12,10 +12,10 @@ import 'package:hirevire_app/common/widgets/text_field.dart';
 import 'package:hirevire_app/constants/color_constants.dart';
 import 'package:hirevire_app/constants/global_constants.dart';
 import 'package:hirevire_app/constants/image_constants.dart';
-import 'package:hirevire_app/user_interface/controllers/user_onb_controller.dart';
+import 'package:hirevire_app/user_interface/controllers/profile_controller.dart';
 import 'package:hirevire_app/utils/size_util.dart';
 
-class SocialUrlsSection extends GetWidget<UserOnbController> {
+class SocialUrlsSection extends GetWidget<ProfileController> {
   const SocialUrlsSection({super.key});
 
   @override
