+++ lib/user_interface/presentation/profile/sliding_base.dart
@@ -3,9 +3,9 @@ import 'package:get/get.dart';
 import 'package:hirevire_app/common/widgets/animated_progress_indicator.dart';
 import 'package:hirevire_app/common/widgets/padded_container.dart';
 import 'package:hirevire_app/common/widgets/spacing_widget.dart';
-import 'package:hirevire_app/user_interface/controllers/user_onb_controller.dart';
+import 'package:hirevire_app/user_interface/controllers/profile_controller.dart';
 
-class SlidingBase extends GetWidget<UserOnbController> {
+class SlidingBase extends GetWidget<ProfileController> {
   const SlidingBase({super.key});
 
   @override
